pyproject.toml
[project]
name = "bq-extraction-demo"
version = "0.1.0"
description = "BigQuery extraction demo tooling"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bq_extraction_demo"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
